fix: rename typo in file
